Split api tests into separate files based on the endpoint.
